Removing Proxy features  (#1143)
