+++ Block/Adminhtml/Category/Merchandising.php
@@ -4,7 +4,6 @@ namespace Algolia\AlgoliaSearch\Block\Adminhtml\Category;
 
 use Algolia\AlgoliaSearch\Helper\ConfigHelper;
 use Algolia\AlgoliaSearch\Helper\Data;
-use Algolia\AlgoliaSearch\Helper\ProxyHelper;
 use Magento\Backend\Block\Template\Context;
 use Magento\Catalog\Model\Category;
 use Magento\Framework\Registry;
@@ -23,9 +22,6 @@ class Merchandising extends \Magento\Backend\Block\Template
     /** @var Data */
     private $coreHelper;
 
-    /** @var ProxyHelper */
-    private $proxyHelper;
-
     /** @var \Magento\Store\Model\StoreManagerInterface */
     private $storeManager;
 
@@ -34,7 +30,6 @@ class Merchandising extends \Magento\Backend\Block\Template
      * @param Registry $registry
      * @param ConfigHelper $configHelper
      * @param Data $coreHelper
-     * @param ProxyHelper $proxyHelper
      * @param array $data
      */
     public function __construct(
@@ -42,13 +37,11 @@ class Merchandising extends \Magento\Backend\Block\Template
         Registry $registry,
         ConfigHelper $configHelper,
         Data $coreHelper,
-        ProxyHelper $proxyHelper,
         array $data = []
     ) {
         $this->registry = $registry;
         $this->configHelper = $configHelper;
         $this->coreHelper = $coreHelper;
-        $this->proxyHelper = $proxyHelper;
         $this->storeManager = $context->getStoreManager();
 
         parent::__construct($context, $data);
@@ -87,20 +80,6 @@ class Merchandising extends \Magento\Backend\Block\Template
     public function getCoreHelper()
     {
         return $this->coreHelper;
-    }
-
-    /** @return bool */
-    public function isQueryRulesEnabled()
-    {
-        $info = $this->proxyHelper->getInfo(ProxyHelper::INFO_TYPE_QUERY_RULES);
-
-        // In case the call to API proxy fails,
-        // be "nice" and return true
-        if ($info && array_key_exists('query_rules', $info)) {
-            return $info['query_rules'];
-        }
-
-        return true;
     }
 
     /**
