+++ Block/Adminhtml/LandingPage/SearchConfiguration.php
@@ -4,7 +4,6 @@ namespace Algolia\AlgoliaSearch\Block\Adminhtml\LandingPage;
 
 use Algolia\AlgoliaSearch\Helper\ConfigHelper;
 use Algolia\AlgoliaSearch\Helper\Data;
-use Algolia\AlgoliaSearch\Helper\ProxyHelper;
 use Algolia\AlgoliaSearch\Model\LandingPage;
 use Magento\Backend\Block\Template\Context;
 use Magento\Framework\Registry;
@@ -23,9 +22,6 @@ class SearchConfiguration extends \Magento\Backend\Block\Template
     /** @var Data */
     private $coreHelper;
 
-    /** @var ProxyHelper */
-    protected $proxyHelper;
-
     /** @var int */
     protected $planLevel;
 
@@ -34,7 +30,6 @@ class SearchConfiguration extends \Magento\Backend\Block\Template
      * @param Registry $registry
      * @param ConfigHelper $configHelper
      * @param Data $coreHelper
-     * @param ProxyHelper $proxyHelper
      * @param array $data
      */
     public function __construct(
@@ -42,13 +37,11 @@ class SearchConfiguration extends \Magento\Backend\Block\Template
         Registry $registry,
         ConfigHelper $configHelper,
         Data $coreHelper,
-        ProxyHelper $proxyHelper,
         array $data = []
     ) {
         $this->registry = $registry;
         $this->configHelper = $configHelper;
         $this->coreHelper = $coreHelper;
-        $this->proxyHelper = $proxyHelper;
 
         parent::__construct($context, $data);
     }
@@ -69,21 +62,5 @@ class SearchConfiguration extends \Magento\Backend\Block\Template
     public function getCoreHelper()
     {
         return $this->coreHelper;
-    }
-
-    /** @return int */
-    public function getPlanLevel()
-    {
-        if ($this->planLevel == null) {
-            $planLevel = 1;
-            $planLevelInfo = $this->proxyHelper->getInfo(ProxyHelper::INFO_TYPE_PLAN_LEVEL);
-
-            if (isset($planLevelInfo['plan_level'])) {
-                $planLevel = (int) $planLevelInfo['plan_level'];
-            }
-            $this->planLevel = $planLevel;
-        }
-
-        return $this->planLevel;
     }
 }
