+++ Helper/AnalyticsHelper.php
@@ -23,9 +23,6 @@ class AnalyticsHelper
     /** @var IndexEntityDataProvider */
     private $entityHelper;
 
-    /** @var ProxyHelper */
-    private $proxyHelper;
-
     /** @var Logger */
     private $logger;
 
@@ -57,21 +54,18 @@ class AnalyticsHelper
      * @param AlgoliaHelper $algoliaHelper
      * @param ConfigHelper $configHelper
      * @param IndexEntityDataProvider $entityHelper
-     * @param ProxyHelper $proxyHelper
      * @param Logger $logger
      */
     public function __construct(
         AlgoliaHelper $algoliaHelper,
         ConfigHelper $configHelper,
         IndexEntityDataProvider $entityHelper,
-        ProxyHelper $proxyHelper,
         Logger $logger
     ) {
         $this->algoliaHelper = $algoliaHelper;
         $this->configHelper = $configHelper;
 
         $this->entityHelper = $entityHelper;
-        $this->proxyHelper = $proxyHelper;
 
         $this->logger = $logger;
     }
@@ -305,25 +299,9 @@ class AnalyticsHelper
         return $conversion && isset($conversion['dates']) ? $conversion['dates'] : [];
     }
 
-    /**
-     * Client Data Check
-     *
-     * @return mixed
-     */
-    public function getClientData()
-    {
-        if (!$this->clientData) {
-            $this->clientData = $this->proxyHelper->getInfo(ProxyHelper::INFO_TYPE_ANALYTICS);
-        }
-
-        return $this->clientData;
-    }
-
     public function isAnalyticsApiEnabled()
     {
-        $clientData = $this->getClientData();
-
-        return (bool) $clientData && isset($clientData['analytics_api']) ? $clientData['analytics_api'] : 1;
+        return true;
     }
 
     public function isClickAnalyticsEnabled()
@@ -332,9 +310,7 @@ class AnalyticsHelper
             return false;
         }
 
-        $clientData = $this->getClientData();
-
-        return (bool) $clientData && isset($clientData['click_analytics']) ? $clientData['click_analytics'] : 1;
+        return true;
     }
 
     /**
