+++ Model/Observer/SaveSettings.php
@@ -3,7 +3,6 @@
 namespace Algolia\AlgoliaSearch\Model\Observer;
 
 use Algolia\AlgoliaSearch\Exceptions\AlgoliaException;
-use Algolia\AlgoliaSearch\Model\ConfigurationTracker;
 use Algolia\AlgoliaSearch\Model\IndicesConfigurator;
 use Magento\Framework\Event\Observer;
 use Magento\Framework\Event\ObserverInterface;
@@ -17,17 +16,12 @@ class SaveSettings implements ObserverInterface
     /** @var IndicesConfigurator */
     private $indicesConfigurator;
 
-    /** @var ConfigurationTracker */
-    private $configurationTracker;
-
     public function __construct(
         StoreManagerInterface $storeManager,
-        IndicesConfigurator $indicesConfigurator,
-        ConfigurationTracker $configurationTracker
+        IndicesConfigurator $indicesConfigurator
     ) {
         $this->storeManager = $storeManager;
         $this->indicesConfigurator = $indicesConfigurator;
-        $this->configurationTracker = $configurationTracker;
     }
 
     /**
@@ -41,7 +35,6 @@ class SaveSettings implements ObserverInterface
 
         foreach ($storeIds as $storeId) {
             $this->indicesConfigurator->saveConfigurationToAlgolia($storeId);
-            $this->configurationTracker->trackConfiguration($storeId);
         }
     }
 }
