--- view/adminhtml/layout/algolia_algoliasearch_handle_query_access_denied.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0"?>
-<page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" layout="admin-1column" xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
-    <head>
-        <css src="Algolia_AlgoliaSearch::css/landing-page.css"/>
-        <title>Algolia | Query Merchandiser</title>
-    </head>
-    <body>
-        <referenceContainer name="content">
-            <block class="Magento\Backend\Block\Denied" name="content.denied" template="Algolia_AlgoliaSearch::query/access_denied.phtml"/>
-        </referenceContainer>
-    </body>
-</page>
