+++ view/adminhtml/layout/algolia_algoliasearch_merchandising_index.xml
@@ -7,14 +7,7 @@
     </head>
     <body>
         <referenceContainer name="content">
-
-            <block class="Magento\Backend\Block\Template" name="merchandising_page" template="Algolia_AlgoliaSearch::merchandising/page.phtml" >
-                <arguments>
-                    <argument name="view_model" xsi:type="object">Algolia\AlgoliaSearch\ViewModel\Adminhtml\Merchandising\Page</argument>
-                </arguments>
-                <block class="Magento\Backend\Block\Template" name="upsell" template="Algolia_AlgoliaSearch::ui/upsell-plan.phtml" ></block>
-            </block>
-
+            <block class="Magento\Backend\Block\Template" name="merchandising_page" template="Algolia_AlgoliaSearch::merchandising/page.phtml" />
         </referenceContainer>
     </body>
 </page>
