+++ view/adminhtml/layout/algolia_algoliasearch_support_index.xml
@@ -12,11 +12,6 @@
                 <arguments>
                     <argument name="view_model" xsi:type="object">Algolia\AlgoliaSearch\ViewModel\Adminhtml\Support\Overview</argument>
                 </arguments>
-                <block class="Magento\Backend\Block\Template" name="support_contact" template="Algolia_AlgoliaSearch::support/contact.phtml">
-                    <arguments>
-                        <argument name="view_model" xsi:type="object">Algolia\AlgoliaSearch\ViewModel\Adminhtml\Support\Contact</argument>
-                    </arguments>
-                </block>
             </block>
         </referenceContainer>
     </body>
