--- view/adminhtml/templates/query/access_denied.phtml
@@ -1,21 +1,0 @@
-<?php
-/**
- * @see \Magento\Backend\Block\Denied
- */
-$emptyStateIllustration = $block->getViewFileUrl('Algolia_AlgoliaSearch::images/illu-querymerchandiser.svg');
-$bookIllustration = $block->getViewFileUrl('Algolia_AlgoliaSearch::images/icon-book.svg');
-$starsIllustration = $block->getViewFileUrl('Algolia_AlgoliaSearch::images/icon-stars.svg');
-?>
-<hr class="access-denied-hr"/>
-<div class="access-denied-page algolia-denied-page">
-    <span><img src="<?php echo $block->escapeUrl($emptyStateIllustration) ?>" /></span>
-    <div class="algolia-suggestions-header">
-        <img src="<?php echo $block->escapeUrl($bookIllustration) ?>" class="book" />
-        <p>Easily promote or demote products for a search. Based on Algolia's Query rules, the Query Merchandiser allows you to enchance the ranking behaviour for specific search queries.</p>
-        <p>Merchandising a query will create a new query rule (which are part of your Algolia plan). To find out more about the Query Merchandiser, please see our <a href="https://www.algolia.com/doc/integration/magento-2/merchandising/query-merchandiser/?utm_source=magento&utm_medium=extension&utm_campaign=magento_2&utm_term=shop-owner&utm_content=doc-link" target="_blank">documentation</a>.</p>
-    </div>
-    <div class="algolia-suggestions-header algolia-suggestions-header-wbg">
-        <img src="<?php echo $block->escapeUrl($starsIllustration) ?>" class="stars" />
-        <p>To get access to this Algolia feature, please consider <a href="https://www.algolia.com/pricing" target="_blank" class="algolia-upgrade-plan">upgrading to a higher plan</a>.</p>
-    </div>
-</div>
